final-pjt-back/requirements.txt
anyio==3.3.0
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.1
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
colorama==0.4.4
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.6
django-cors-headers==3.10.0
django-extensions==3.1.3
django-seed==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
Faker==8.10.3
fett==0.3.2
idna==3.2
ipykernel==6.0.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.10.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.10.1
jupyterlab==3.1.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
jupyterlab-widgets==1.0.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.0
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
pymongo==3.11.4
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-jsonrpc-server==0.3.4
pytz==2021.1
pywin32==301
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.26.0
requests-unixsocket==0.2.0
rstcheck==3.3.1
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.7.1
six==1.16.0
sklearn==0.0
sniffio==1.2.0
snooty-lextudio==1.11.4.dev0
soupsieve==2.2.1
sqlparse==0.4.1
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
toposort==1.7
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.6
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
widgetsnbextension==3.5.1
